Fix last 4 warnings known.

+++ sdk/include/opentelemetry/sdk/metrics/exemplar/histogram_exemplar_reservoir.h
@@ -42,18 +42,18 @@ public:
   public:
     HistogramCellSelector(const std::vector<double> &boundaries) : boundaries_(boundaries) {}
 
-    int ReservoirCellIndexFor(const std::vector<ReservoirCell> &cells,
-                              long value,
-                              const MetricAttributes &attributes,
-                              const opentelemetry::context::Context &context) override
+    size_t ReservoirCellIndexFor(const std::vector<ReservoirCell> &cells,
+                                 long value,
+                                 const MetricAttributes &attributes,
+                                 const opentelemetry::context::Context &context) override
     {
       return ReservoirCellIndexFor(cells, (double)value, attributes, context);
     }
 
-    int ReservoirCellIndexFor(const std::vector<ReservoirCell> & /* cells */,
-                              double value,
-                              const MetricAttributes & /* attributes */,
-                              const opentelemetry::context::Context & /* context */) override
+    size_t ReservoirCellIndexFor(const std::vector<ReservoirCell> & /* cells */,
+                                 double value,
+                                 const MetricAttributes & /* attributes */,
+                                 const opentelemetry::context::Context & /* context */) override
     {
       for (size_t i = 0; i < boundaries_.size(); ++i)
       {
